Refactor cli.py into cli subfolder
