+++ src/agendable/cli/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from agendable.cli.main import main
+
+__all__ = ["main"]
